+++ Plugins/MongoDBDriverPlugin/MongoDBPlugin.swift
@@ -97,6 +97,7 @@ final class MongoDBPlugin: NSObject, TableProPlugin, DriverPlugin {
     static let editorLanguage: EditorLanguage = .javascript
     static let supportsForeignKeys = false
     static let supportsSchemaEditing = false
+    static let supportsRenameView = false
     static let systemDatabaseNames: [String] = ["admin", "local", "config"]
     static let tableEntityName = "Collections"
     static let supportsForeignKeyDisable = false